[Medical History View] - Fixed Dropdown style
