+++ lhapp/src/components/Inputs/Dropdown.css
@@ -1,0 +1,70 @@
+.dropdown-input{
+    width: 100%;
+    height: 25px;
+    margin: 10px 0;
+    border-radius: 10px;
+    padding: 5px;
+    box-shadow: 0 0 0 1.5pt var(--gray2);
+    font-size: var(--sFont);
+}
+
+.dropdown-input span{
+    font-size: var(--sFont);
+    font-weight: normal ;
+}
+
+.dropdown-input .dropdown-button{
+    z-index: 1000;
+    width: inherit;
+    cursor: pointer;
+    height: 25px;
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    background-color: none !important;
+    margin-bottom: 10px;
+}
+
+.dropdown-input .arrowMenu{
+    width: 15px;
+    margin: 0 10px;
+    transform: rotateZ(-180deg);
+    transition: all .3s ease-in-out;
+}
+
+.dropdown-input .arrowMenu.active{
+    transform: rotateZ(0deg);
+}
+
+.dropdown-input .dropdown-button > span{
+    width: 100%;
+    font-size: 18px;
+    font-weight: bold;
+}
+
+.dropdown-menu{
+    display: none;
+    box-shadow: 0 0 7px rgba(0, 0, 0, 0.1);
+}
+
+.dropdown-menu.active{
+    display: block;
+    position: relative;
+    z-index: 1000;
+    list-style: none;
+    background-color: white;
+    padding: 0;
+    margin: 0;
+    border-radius: 10px;
+}
+
+.dropdown-menu.active li{
+    padding: 10px 25px;
+    font-size: 20px;
+}
+
+.dropdown-menu.active li:hover{
+    cursor: pointer;
+    background-color: var(--bgColor);
+    border-radius: 10px;
+}
